requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==1.10.17
xmltodict==0.13.0
python-multipart==0.0.12
PyJWT==2.9.0